refactor(core): lift protect managers into unifi_core.protect.managers
10 manager modules moved with backwards-compatible shims at the old paths.
Cross-manager and lifted-shared imports rewritten. Shims removed in Task 17.

Updated two unit tests to patch the new unifi_core.protect.managers.* module
paths since the patched names (ProtectApiClient, retry_with_backoff,
asyncio.sleep) are now resolved in the core module's namespace.
